Fix typeahead matcher so it works with spaces
The default matcher breaks (i.e. no suggestions are shown) if user enters text with spaces because of the HTML-markup inserted in the item text. I'm not sure the matcher should get the highlighted version of the item so I made this workaround. Probably, there could be a better fix.